Move the rewritten functions back to where they came from.
