+++ cms/menu.py
@@ -5,7 +5,7 @@ from cms.utils import get_language_from_request
 from cms.utils.i18n import get_fallback_languages, hide_untranslated
 from cms.utils.page_resolver import get_page_queryset
 from cms.utils.moderator import get_title_queryset
-from cms.utils import permissions
+from cms.utils.permissions import load_view_restrictions, has_global_page_permission
 from cms.utils.plugins import current_site
 from menus.base import Menu, NavigationNode, Modifier
 from menus.menu_pool import menu_pool
@@ -14,9 +14,79 @@ from django.conf import settings
 from django.utils.translation import get_language
 
 
+def get_visible_page_objects(request, pages, site=None):
+    """
+     This code is basically a many-pages-at-once version of
+     Page.has_view_permission.
+     pages contains all published pages
+     check if there is ANY restriction
+     that needs a permission page visibility calculation
+    """
+    can_see_unrestricted = settings.CMS_PUBLIC_FOR == 'all' or (
+        settings.CMS_PUBLIC_FOR == 'staff' and request.user.is_staff)
+    is_auth_user = request.user.is_authenticated()
+
+    restricted_pages = load_view_restrictions(request, pages)
+    if not restricted_pages:
+        if can_see_unrestricted:
+            return pages
+        elif not is_auth_user:
+            return [] # Unauth user can't acquire global or user perm to see pages
+
+    if settings.CMS_PERMISSION and not site:
+        site = current_site(request) # avoid one extra query when possible
+    if has_global_page_permission(request, site, can_view=True):
+        return pages
+
+    def has_global_perm():
+        if has_global_perm.cache < 0:
+            has_global_perm.cache = bool(request.user.has_perm('cms.view_page'))
+        return bool(has_global_perm.cache)
+    has_global_perm.cache = -1
+
+    def has_permission_membership(page_id):
+        """
+        PagePermission user group membership tests
+        """
+        user_pk = request.user.pk
+        for perm in restricted_pages[page_id]:
+            if perm.user_id == user_pk:
+                return True
+            if not perm.group_id:
+                continue
+            if has_permission_membership.user_groups is None:
+                has_permission_membership.user_groups = request.user.groups.all().values_list('pk', flat=True)
+            if perm.group_id in has_permission_membership.user_groups:
+                return True
+        return False
+    has_permission_membership.user_groups = None
+
+    visible_pages = []
+    for page in pages:
+        to_add = False
+        page_id = page.pk
+        is_restricted = page_id in restricted_pages
+        # restricted_pages contains as key any page.pk that is
+        # affected by a permission grant_on
+        if not is_restricted and can_see_unrestricted:
+            to_add = True
+        elif is_auth_user:
+            # setting based handling of unrestricted pages
+            # check group and user memberships to restricted pages
+            if is_restricted and has_permission_membership(page_id):
+                to_add = True
+            elif has_global_perm():
+                to_add = True
+
+        if to_add:
+            visible_pages.append(page)
+
+    return visible_pages
+
+
 def get_visible_pages(request, pages, site=None):
     """Returns the IDs of all visible pages"""
-    pages = permissions.get_visible_pages(request, pages, site)
+    pages = get_visible_page_objects(request, pages, site)
     return [page.pk for page in pages]
